refactor: backendの初期化処理と設定を集約する
Prisma と Firebase Admin の初期化が複数箇所に分散していた問題を解消する。

- config/firebase.ts を新規作成し Firebase Admin SDK 初期化を集約
  - auth.middleware.ts の堅牢な初期化ロジック（base64/\n対応, ADCフォールバック）を移植
- config/db.ts を Prisma 唯一の初期化ファイルとして整備
  - dotenv ロードを追加（.env.local 優先）
  - 環境に応じたログレベル設定（本番: error のみ）
- config/prisma.ts を削除（config/db.ts に統合）
- app.ts から dotenv ロードと Firebase 初期化ブロックを削除
- auth.middleware.ts から Firebase 初期化ブロックを削除し config/firebase を利用
- 全 Accessor・questJoinService の import を config/prisma → config/db に統一

closes #53
