+++ apps/backend/src/config/env.ts
@@ -1,0 +1,5 @@
+import * as dotenv from "dotenv";
+import { join } from "node:path";
+
+dotenv.config({ path: join(process.cwd(), ".env.local") });
+dotenv.config();
