+++ apps/backend/src/config/firebase.ts
@@ -1,0 +1,57 @@
+import "./env"; // 環境変数ロード（単一エントリポイント）
+import admin from "firebase-admin";
+
+if (!admin.apps.length) {
+  const projectId = process.env.FIREBASE_PROJECT_ID ?? "";
+  const clientEmail = process.env.FIREBASE_CLIENT_EMAIL ?? "";
+  const rawPrivateKey = process.env.FIREBASE_PRIVATE_KEY ?? "";
+
+  let privateKey: string | undefined;
+  if (rawPrivateKey) {
+    try {
+      const maybeDecoded = Buffer.from(rawPrivateKey, "base64").toString("utf8");
+      if (
+        maybeDecoded.includes("-----BEGIN") &&
+        maybeDecoded.includes("PRIVATE KEY-----")
+      ) {
+        privateKey = maybeDecoded;
+      } else {
+        privateKey = rawPrivateKey.replace(/\\n/g, "\n");
+      }
+    } catch {
+      privateKey = rawPrivateKey.replace(/\\n/g, "\n");
+    }
+  }
+
+  try {
+    if (projectId && clientEmail && privateKey) {
+      admin.initializeApp({
+        credential: admin.credential.cert({ projectId, clientEmail, privateKey }),
+      });
+      console.log("[firebase] Firebase Admin SDK initialized with service account credentials");
+    } else {
+      admin.initializeApp({
+        credential: admin.credential.applicationDefault(),
+      });
+      console.warn(
+        "[firebase] Using Application Default Credentials. Set FIREBASE_PROJECT_ID, FIREBASE_CLIENT_EMAIL, FIREBASE_PRIVATE_KEY to use service account cert."
+      );
+    }
+  } catch (e) {
+    console.error("[firebase] Firebase Admin SDK initialization failed:", e);
+    // サービスアカウント認証に失敗した場合は ADC にフォールバック
+    try {
+      if (!admin.apps.length) {
+        admin.initializeApp({
+          credential: admin.credential.applicationDefault(),
+        });
+        console.warn("[firebase] Falling back to Application Default Credentials after cert failure.");
+      }
+    } catch (fallbackError) {
+      console.error("[firebase] ADC fallback also failed:", fallbackError);
+      console.warn("[firebase] Firebase Admin SDK initialization failed. Authentication will not work properly.");
+    }
+  }
+}
+
+export default admin;
