+++ apps/backend/src/dataAccessor/dbAccessor/Quest.ts
@@ -1,4 +1,4 @@
-import prisma from "../../config/prisma";
+import { prisma } from "../../config/db";
 import { Quest, QuestParticipant, Reward, User } from "@prisma/client";
 
 /**
