+++ apps/backend/src/dataAccessor/dbAccessor/QuestParticipant.ts
@@ -1,4 +1,4 @@
-import prisma from "../../config/prisma";
+import { prisma } from "../../config/db";
 import { QuestParticipant } from "@prisma/client";
 
 /**
