+++ apps/backend/src/dataAccessor/dbAccessor/User.ts
@@ -1,4 +1,4 @@
-import prisma from "../../config/prisma";
+import { prisma } from "../../config/db";
 import { User } from "@prisma/client";
 
 /**
